Reworked outgoing pull replication
It now works more like the discard pass does and no longer does a
preceding scan count.

+++ stats.go
@@ -31,7 +31,7 @@ type valueStoreStats struct {
 	pathtoc                    string
 	workers                    int
 	tombstoneDiscardInterval   int
-	outPullReplicationWorkers  int
+	outPullReplicationWorkers  uint64
 	outPullReplicationInterval int
 	outPushReplicationWorkers  int
 	outPushReplicationInterval int
